Updated GUI, API, and Add Container
created page for http 400 errors,
Made api route to create new Containers.
When creating new containers the filepath.isAbs will be used to check where the container will be stored. If the path is absolute that location will be used if not the path will be added to the working directory path and stored there ( /DatabaseDepot/* ). Renamed the default container to default.
TODO - Create a config file used to store and load contains when the application starts and exists